# postman/openapi.yaml
openapi: 3.1.0
info:
  title: ecommerce
  version: 1.0.0
servers:
  - url: https://{{host}}
  - url: https://ecommerce-tp2e.onrender.com
paths:
...
  /upload:
    post:
      summary: '{{host}}/upload'
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
...
  /user/preferences/save:
    post:
      summary: '{{host}}/user/preferences/save'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                preferences:
                  product:
                    is_display_name: true
                    is_display_price: true
                    is_display_sku: true
                    is_display_stock: false
                  category:
                    is_display_name: true
                    is_display_description: true
                  order:
                    is_display_orderId: true
                    is_display_status: true
                    is_display_totalAmount: true
...
  /import-csv:
    post:
      summary: '{{host}}/import-csv'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
...